remove duplicate outdated unparserfilter class
